added folders to sort components
