+++ src/App.vue
@@ -11,8 +11,8 @@
 </template>
 
 <script setup>
-import Footer from "@/components/Footer.vue";
-import NavBar from "@/components/NavBar.vue";
+import Footer from "@/components/nav/Footer.vue";
+import NavBar from "@/components/nav/NavBar.vue";
 import { useRouteStore } from "@/store/index.js"; //
 import { storeToRefs } from "pinia";
 import { ref, onMounted } from "vue";
