--- src/components/RouteCard.vue
@@ -1,34 +1,0 @@
-<template>
-  <v-card flat class="text-xs-center ma-3">
-    <v-container>
-      <CardMap :route="route" />
-      <v-card-title>{{ route.name }}</v-card-title>
-      <v-card-text>{{ route.length }} miles</v-card-text>
-      <v-card-text>{{ route.elevation }} elevation</v-card-text>
-
-      <v-card-actions>
-        <v-btn router :to="link"> View Route </v-btn>
-      </v-card-actions>
-    </v-container>
-  </v-card>
-</template>
-
-<script>
-import CardMap from "@/components/CardMap.vue";
-export default {
-  props: {
-    route: null, //route inputed by parent component
-    link: null,
-  },
-  data() {
-    return {
-      link: "/route/" + this.route.id,
-    };
-  },
-  components: {
-    CardMap,
-  },
-};
-</script>
-
-<style></style>
